Renamed the directives module to csp as again this is clearer for the domain.
